Fixed a bug with m_destroy in FbTk::FbWindow.
If one creates a FbWindow using the protected constructor that takes a
Window as an argument and then makes a copy of it, m_destroy will be set
to true in the copy. As a result, whenever that FbWindow object is
destroyed, the class the tries to destroy the corresponding X window.
Since the whole purpose of that constructor is to create a FbWindow
object that refers to some existing X window (i.e. it does not control
its lifetime), this is a bug.

With this change, the code copies the value of m_destroy variable when
a copy of a FbWindow object is made, fixing the above behaviour.