Fixed the GLXEW initialization issue.
GLXEW is GLEW for GLX extensions. Using it anywhere zeroes the pointers
to GLX functions, making them unusable. The problem is that we have to
call glewInit to initialize them, but for that we need to create an
OpenGL context, but for that we need to initialize appropriate functions
with glewInit... you get the idea.

This commit manually initializes a small amount of GLX functions, just
enough to create a context, before GLEW takes care of the rest.