Added a new timer class.
This new class allows to track time intervals much easier, compared to
FbTk::Timer.

It is also more accurate, since FbTk::Timer constantly refreshes the
start time when it is told to perform something on an interval basis and
it does that *after* the action in question is performed (so if you take
something that takes up non-trivial amount of time, like rendering, you
will lose time "ticks").

The code has been updated to make use of this new timer.