Optimized shape handling in OpenGL renderer.
Instead of relying on a slow X server operation to add the appropriate
alpha values to the window's content texture, we now create a separate
OpenGL texture and apply the transparency in the fragment shader. This
simple change resulted in a noticeable performance increase and (at
least on my machine) the compositing is no longer intrusive in OpenGL
rendering mode.

There are several code inconsistencies that I will address later.

Other changes:
> Added a macro to suppress the unused parameter compiler warning.
> Moved the pixmap to texture conversion code to a separate function for
  public use.
> Improved sleeping behaviour in the compositor's event loop.
> Comment additions in the rendering code. It should clearer now.
> Appropriate shader changes.
> Other small tweaks.