Implemented OpenGL texture splitting. I think.
At least on my machine, I can reduce the maximum texture size to 512 and
not see any difference in the picture quality (no seams or anything), so
I assume that the partitioner works.

Now, this code needs to be refactored badly. I did some copy-pasting to
get this working because it is late and so that others can test the
implementation. API in particular needs to be cleaned up and the there
is some copypasta.

Also, there has to be a more efficient method of accessing sub-pixmaps
than XCopyArea. This causes the largest performance drops.