Refactored yesterday's work.
Changes:
> Moved the texture partitioner to a separate set of files. Also added
  publicly accessible functions to partition some rectangle and
  calculate what borders is a given partition adjacent to.
> Content access API in OpenGLWindow is actually sane now.
> Completely refactored OpenGLWindow's updateWindowPos function.
> Refactored window rendering code in OpenGLScreen. Now the plugins are
  aware of which partition is being rendered.
> Added an overloaded version of toOpenGLCoordinates.
> Added a function to the buffer wrapper that allows it to easily buffer
  rectangle's OpenGL coordinates.
> Compositor will now sync with the server after rendering a frame.
> Added IndexException for index-related errors (i.e. out of bounds).
> Removed several unused Display* member variables.

There is also the issue of how the partitioner splits the texture. At
the moment what it does is it takes the content pixmap, takes its
subpixmaps using XCopyArea and uses those to create partition textures.
There are two problems with this:
1) It is CPU intensive.
2) The application keeps updating that pixmap as it is being copied.
Because of the point #2, you might see seams in "dynamic" applications,
such as glxgears. I will find a solution for this, but for now the
compositor should at least run in OpenGL mode.